# docs/swagger.yaml
definitions:
  dto.Pageable:
    properties:
      offset:
        type: integer
      page:
        type: integer
      size:
        type: integer
    type: object
  dto.Sortable:
    properties:
      direction:
        type: string
      field:
        type: string
    type: object
...
  request.EmployeeTrainingRequest:
    properties:
      id:
        type: integer
      karyawan:
        $ref: '#/definitions/request.IdRequest'
      tanggal:
        type: string
      training:
        $ref: '#/definitions/request.IdRequest'
    required:
    - tanggal
    type: object
  request.IdRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  request.TrainingRequest:
    properties:
      id:
        type: integer
      pengajar:
        type: string
      tema:
        type: string
    required:
    - pengajar
    - tema
    type: object
  response.ApiResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  response.PaginationData:
    properties:
      content: {}
      emptyPage:
        type: boolean
      firstPage:
        type: boolean
      lastPage:
        type: boolean
      numberOfElements:
        type: integer
      pageable:
        $ref: '#/definitions/dto.Pageable'
      sortable:
        $ref: '#/definitions/dto.Sortable'
      totalElements:
        type: integer
      totalPages:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /v1/idstar/karyawan-training/{id}:
    get:
      consumes:
      - application/json
      operationId: GetEmployeeTrainingById
      parameters:
      - description: Training ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Karyawan Training By Id
      tags:
      - karyawan-training
  /v1/idstar/karyawan-training/delete:
    delete:
      consumes:
      - application/json
      operationId: DeleteEmployeeTraining
      parameters:
      - description: Delete Karyawan Training Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.IdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Karyawan Training
      tags:
      - karyawan-training
  /v1/idstar/karyawan-training/list:
    get:
      consumes:
      - application/json
      operationId: GetEmployeeTrainingList
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Size
        in: query
        name: size
        type: integer
      - description: Field
        in: query
        name: field
        type: string
      - description: Direction
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginationData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Karyawan Training List
      tags:
      - karyawan-training
  /v1/idstar/karyawan-training/save:
    post:
      consumes:
      - application/json
      operationId: CreateEmployeeTraining
      parameters:
      - description: Create Karyawan Training Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.EmployeeTrainingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Karyawan Training
      tags:
      - karyawan-training
  /v1/idstar/karyawan-training/update:
    put:
      consumes:
      - application/json
      operationId: UpdateEmployeeTraining
      parameters:
      - description: Update Karyawan Training Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.EmployeeTrainingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Karyawan Training
      tags:
      - karyawan-training
...
  /v1/idstar/training/{id}:
    get:
      consumes:
      - application/json
      operationId: GetTrainingById
      parameters:
      - description: Training ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Training By Id
      tags:
      - training
  /v1/idstar/training/delete:
    delete:
      consumes:
      - application/json
      operationId: DeleteTraining
      parameters:
      - description: Delete Training Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.IdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Training
      tags:
      - training
  /v1/idstar/training/list:
    get:
      consumes:
      - application/json
      operationId: GetTrainingList
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Size
        in: query
        name: size
        type: integer
      - description: Field
        in: query
        name: field
        type: string
      - description: Direction
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginationData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Training List
      tags:
      - training
  /v1/idstar/training/save:
    post:
      consumes:
      - application/json
      operationId: CreateTraining
      parameters:
      - description: Create Training Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.TrainingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Training
      tags:
      - training
  /v1/idstar/training/update:
    put:
      consumes:
      - application/json
      operationId: UpdateTraining
      parameters:
      - description: Update Training Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.TrainingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Training
      tags:
      - training
...
swagger: "2.0"